+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 1.14.2 - 2025-03-xx
+## Version 1.14.2 - 2025-03-04
 
 ### Fixed
 
 - Specific logger component for http proxy middleware
 - Expose `count` in response for `Atom (XML)` format
+- Option `excludeNonSelectedKeys: true` to exclude non-selected keys from response (OData V4 auto-includes keys)
 
 ## Version 1.14.1 - 2025-02-03